New: Paths for switch from all three staring positions and cases (6) (tested on programming bot) Auto selection from SmartDashboard and game data (LRL, RLR, etc.)
Fixed bug with switching between TeleOp and Auto where the robot is
moved and the NavX reference angle is not updated.

I learned that the MXP power from the roborio occasionally loses power
when motors and other actuators draw too much current, and after adding
a USB cable to the NavX and RoboRIO, this fixed the issue. I read that
some other teams were having problems with this in the past, and that
was the solution.

Videos of all paths have been posted in Discord in the #programming
channel